pyproject.toml
[tool.poetry]
name = "django-jackal"
version = "2.4.0b2"
description = "Boilerplate for Django and Django REST Framework"
authors = ["joyongjin <wnrhd114@gmail.com>"]
license = "MIT"
readme = 'README.rst'
repository = 'https://github.com/joyongjin/jackal'
homepage = 'https://github.com/joyongjin/jackal'
packages = [
    { include = 'jackal' }
]
exclude = ['test*', '.*']

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.2"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
